requirements.txt
django==4.1.5
pillow==9.4.0
django-environ==0.9.0
psycopg2==2.9.5
black==23.1.0
djangorestframework==3.14.0

